Add Event mappings for admin and public DTOs
This commit introduces new mappings for the `Event` entity to various Data Transfer Objects (DTOs) for both admin and public views. It includes mappings for `EventAdminResponseDTO`, `EventAdminCreateDTO`, `EventAdminUpdateDTO`, `EventPublicListDTO`, and `EventPublicDetailsDTO`. The mappings specify how properties from the `Event` entity are mapped to the DTOs, including the omission of certain properties and the assignment of default values.